fix(ios): keep the keyboard off <html> and wait for the web view resize (#9779)
Opening the add-task bar on iOS still hitched for a second or more with a
long task list, and the bar still shot up and dropped back some of the time,
after #9809.

Two things were left. Capacitor's `resize: 'native'` applies the WebView
frame shrink on a timer set to the keyboard animation duration plus 200ms
(Keyboard.m, onKeyboardWillShow), so it lands ~200ms *after*
`keyboardDidShow`. Our settle logic ran at didShow, found an unshrunk
viewport, concluded the web view would never resize, and lifted the bar by a
whole keyboard height — which the late resize then took back: the jump.
And `--keyboard-height` was still written on `<html>` at every show and
hide, on top of the two `--keyboard-overlay-offset` writes that race
produced. Each of those is the document-wide restyle #9809 measured at
~390ms on a 200-task list, three of them per open.

Now:

- After didShow the keyboard waits up to IOS_KEYBOARD_RESIZE_GRACE_MS for
  the web view to shrink; the visualViewport resize settles it early, and
  only a keyboard that really never resizes the web view gets the
  frame-derived offset, after the grace. The scroll-into-view runs once,
  against the final viewport.
- No keyboard variable is written on `<html>` any more. The overlay
  container carries `--visual-viewport-height`, `--keyboard-height` and
  `--keyboard-overlay-offset` (dialogs, sheets and connected overlays
  inherit them; the CDK viewport patch reads from there), and the global
  add-task bar — the one consumer outside that layer — binds
  `--keyboard-overlay-offset` on its own host from a signal. The `:root`
  defaults stay for Android and mobile web, whose tracker is unchanged.

The iOS keyboard handling moves out of GlobalThemeService (30 lines under
the 1200-line cap) into IosKeyboardService, together with its spec.
