+++ src/app/core/theme/global-theme.service.ts
@@ -1,5 +1,4 @@
 import {
-  afterNextRender,
   computed,
   DestroyRef,
   effect,
@@ -53,18 +52,14 @@ import { IS_ANDROID_WEB_VIEW } from '../../util/is-android-web-view';
 import { androidInterface } from '../../features/android/android-interface';
 import { HttpClient } from '@angular/common/http';
 import { CapacitorPlatformService } from '../platform/capacitor-platform.service';
-import { Keyboard, KeyboardInfo, KeyboardPlugin } from '@capacitor/keyboard';
-import { PluginListenerHandle, registerPlugin } from '@capacitor/core';
-import { OverlayContainer } from '@angular/cdk/overlay';
+import { registerPlugin } from '@capacitor/core';
 import { StatusBar, Style } from '@capacitor/status-bar';
 import { SafeArea } from 'capacitor-plugin-safe-area';
 import { patchCdkViewportForSafeArea } from './cdk-safe-area-viewport.util';
 import { LS } from '../persistence/storage-keys.const';
 import { Log, PluginLog } from '../log';
 import { LayoutService } from '../../core-ui/layout/layout.service';
-import { sanitizeIosKeyboardHeight } from './sanitize-ios-keyboard-height.util';
-import { computeIosKeyboardViewportVars } from './ios-keyboard-viewport-vars.util';
-import { createOneShotSettle } from './ios-keyboard-settle.util';
+import { CSS_VAR_KEYBOARD_HEIGHT, IosKeyboardService } from './ios-keyboard.service';
 import { sanitizeSvgIconContent } from '../../util/sanitize-svg-icon.util';
 import { CustomThemeService, getRequiredThemeMode } from './custom-theme.service';
 
@@ -77,17 +72,6 @@ const NavigationBar = registerPlugin<NavigationBarPlugin>('NavigationBar');
 
 export type DarkModeCfg = 'dark' | 'light' | 'system';
 
-/**
- * How long to wait for a `keyboardDid…` event before acting as if it arrived.
- *
- * Comfortably longer than the keyboard animation, which iOS runs at ~250ms.
- * See `createOneShotSettle` for why neither half of the pair may be trusted.
- */
-const IOS_KEYBOARD_SETTLE_FALLBACK_MS = 400;
-
-const CSS_VAR_KEYBOARD_HEIGHT = '--keyboard-height';
-const CSS_VAR_KEYBOARD_OVERLAY_OFFSET = '--keyboard-overlay-offset';
-const CSS_VAR_VISUAL_VIEWPORT_HEIGHT = '--visual-viewport-height';
 const CSS_VAR_SAFE_AREA_TOP = '--safe-area-inset-top';
 const CSS_VAR_SAFE_AREA_BOTTOM = '--safe-area-inset-bottom';
 const CSS_VAR_SAFE_AREA_LEFT = '--safe-area-inset-left';
@@ -212,44 +196,8 @@ export class GlobalThemeService {
   private _environmentInjector = inject(EnvironmentInjector);
   private _destroyRef = inject(DestroyRef);
   private _inputIntentService = inject(InputIntentService);
-  private _overlayContainer = inject(OverlayContainer);
+  private _iosKeyboardService = inject(IosKeyboardService);
   private _hasInitialized = false;
-  private _keyboardListenerHandles: PluginListenerHandle[] = [];
-  private _focusinListener: ((event: FocusEvent) => void) | null = null;
-  private _visualViewportResizeListener: (() => void) | null = null;
-  private _iosKeyboardHeight = 0;
-  private _iosViewportHeightBeforeKeyboard = 0;
-  private _iosViewportChangeRaf: number | null = null;
-  // True only when the plugin reported an implausible keyboard frame (the clamp
-  // had to correct it). Gates the measured-viewport override so well-behaved
-  // keyboards keep their exact pre-existing behaviour (#8778).
-  private _iosKeyboardFrameUnreliable = false;
-  // False until `keyboardDidShow`: the web view may still be resizing around the
-  // keyboard, so any layout derived from the reported frame would be a guess (#9779).
-  private _isIosKeyboardSettled = false;
-  // iOS drops the `did` half of its keyboard animation pairs, so neither the
-  // settle nor the baseline reset may hang on one alone (#9779).
-  private readonly _iosShowSettle = createOneShotSettle(IOS_KEYBOARD_SETTLE_FALLBACK_MS);
-  private readonly _iosHideSettle = createOneShotSettle(IOS_KEYBOARD_SETTLE_FALLBACK_MS);
-  // Last value written per CSS variable, so a repeated write (the keyboard
-  // animation fires many identical visualViewport resizes) costs nothing.
-  private readonly _cssVarCache = new Map<string, string>();
-  // Where --visual-viewport-height goes; see _initIOSKeyboardHandling, which
-  // assigns it before the first update and before any listener is registered.
-  // Definitely assigned rather than nullable: a fallback to <html> here would be
-  // cached by _setCssVar and silently suppress the later write to the real
-  // target, leaving dialogs on the 100vh fallback for the session.
-  private _iosViewportVarTarget!: HTMLElement;
-
-  /**
-   * Height for the app shell while the iOS keyboard is open, as a CSS value, or
-   * null to leave the sizing to the stylesheet. Bound by app.component rather
-   * than published as a custom property: the shell wraps the whole task list,
-   * and a variable it inherits costs a document-wide style recalc on every
-   * frame of the keyboard animation (#9779).
-   */
-  readonly iosShellHeight = signal<string | null>(null);
-
   private _isCustomWindowTitleBarEnabled(): boolean {
     // The main process (main-window.ts) force-disables the custom title bar on
     // GNOME+Wayland because the Window-Controls-Overlay won't render there.
@@ -558,7 +506,7 @@ export class GlobalThemeService {
 
       if (this._platformService.isIOS()) {
         this.document.body.classList.add(BodyClass.isIOS);
-        this._initIOSKeyboardHandling();
+        this._iosKeyboardService.init();
 
         // Add iPad-specific class for tablet optimizations
         if (this._platformService.isIPad()) {
@@ -593,8 +541,9 @@ export class GlobalThemeService {
 
     // VisualViewport keyboard-height tracking covers every non-iOS touch
     // build: Capacitor Android, the legacy F-Droid build, and Android
-    // mobile-web. iOS uses _initIOSKeyboardHandling above; its Capacitor
-    // plugin already drives the same CSS variable and the two would race.
+    // mobile-web. iOS uses IosKeyboardService above; its Capacitor plugin
+    // drives the same CSS variable (on the overlay container) and the two
+    // would race.
     if (IS_TOUCH_ONLY && !this._platformService.isIOS()) {
       this._initVisualViewportKeyboardTracking();
     }
@@ -707,251 +656,6 @@ export class GlobalThemeService {
   }
 
   /**
-   * Initialize iOS keyboard visibility tracking using Capacitor Keyboard plugin.
-   * Adds/removes CSS classes when keyboard shows/hides.
-   */
-  private _initIOSKeyboardHandling(keyboard: KeyboardPlugin = Keyboard): void {
-    // Hide the native iOS accessory bar (prev/next/Done) — no multi-field forms
-    // benefit from it, and Done is redundant with the system dismiss gesture.
-    keyboard.setAccessoryBarVisible({ isVisible: false });
-    // Resolved up front (this creates the container if CDK has not yet needed
-    // it) so a dialog opened while the keyboard is already up finds the
-    // variable in place.
-    this._iosViewportVarTarget = this._overlayContainer.getContainerElement();
-    this._updateIOSKeyboardViewportVars();
-
-    if (window.visualViewport) {
-      this._visualViewportResizeListener = (): void => {
-        this._updateIOSKeyboardViewportVars();
-      };
-      window.visualViewport.addEventListener(
-        'resize',
-        this._visualViewportResizeListener,
-        { passive: true },
-      );
-    }
-
-    keyboard
-      .addListener('keyboardWillShow', (info: KeyboardInfo) => {
-        Log.log('iOS keyboard will show', info);
-        // Switching to the emoji panel or a taller third-party keyboard re-fires
-        // this while the keyboard is already up, and iOS then reports the already
-        // shrunken window — keeping the first measurement is what stops the
-        // keyboard being subtracted twice.
-        const wasKeyboardVisible = this.document.body.classList.contains(
-          BodyClass.isKeyboardVisible,
-        );
-        // The keyboard is coming back, so the pending baseline reset from the
-        // willHide that preceded a focus move must not fire behind it.
-        this._iosHideSettle.cancel();
-        // Also skipped while a baseline is still held: willHide clears the body
-        // class but not the baseline, so a focus move between two fields lands
-        // here with the web view still shrunken. Only keyboardDidHide clears it.
-        if (!wasKeyboardVisible && !this._iosViewportHeightBeforeKeyboard) {
-          this._iosViewportHeightBeforeKeyboard = window.innerHeight;
-        }
-        // Some third-party keyboards (e.g. Sogou) report a bogus near-full-screen
-        // keyboard frame here; clamp it so it can't fling the fixed add-task bar
-        // to the top of the screen (#8778).
-        const referenceHeight =
-          this._iosViewportHeightBeforeKeyboard || window.innerHeight;
-        const keyboardHeight = sanitizeIosKeyboardHeight(
-          info.keyboardHeight,
-          referenceHeight,
-        );
-        // Only a frame the clamp had to correct opts into the measured-viewport
-        // override in _updateIOSKeyboardViewportVars; well-behaved keyboards keep
-        // the exact pre-existing behaviour, so this cannot regress them.
-        this._iosKeyboardFrameUnreliable = keyboardHeight !== info.keyboardHeight;
-        this._iosKeyboardHeight = keyboardHeight;
-        // The show animation is only starting — nothing about the web view's new
-        // size is known yet, see computeIosKeyboardViewportVars. A keyboard that
-        // is already up is not appearing from zero, though: unsettling it there
-        // would drop the fixed bar behind the keyboard until didShow arrives.
-        if (!wasKeyboardVisible) {
-          this._isIosKeyboardSettled = false;
-          this._iosShowSettle.arm(() => this._settleIosKeyboard());
-        }
-        this.document.body.classList.add(BodyClass.isKeyboardVisible);
-        // Set CSS variable for keyboard height to adjust layout
-        this._setCssVar(
-          this.document.documentElement,
-          CSS_VAR_KEYBOARD_HEIGHT,
-          `${keyboardHeight}px`,
-        );
-        this._updateIOSKeyboardViewportVars();
-      })
-      .then((handle) => this._keyboardListenerHandles.push(handle));
-
-    // Use keyboardDidShow for scroll (after animation completes)
-    keyboard
-      .addListener('keyboardDidShow', () => this._iosShowSettle.run())
-      .then((handle) => this._keyboardListenerHandles.push(handle));
-
-    keyboard
-      .addListener('keyboardWillHide', () => {
-        Log.log('iOS keyboard will hide');
-        this._iosKeyboardHeight = 0;
-        this._iosKeyboardFrameUnreliable = false;
-        this._isIosKeyboardSettled = false;
-        this._iosShowSettle.cancel();
-        // _iosViewportHeightBeforeKeyboard deliberately survives this event:
-        // moving focus between two fields fires willHide then willShow with the
-        // web view still shrunken, and re-snapshotting window.innerHeight there
-        // would subtract the keyboard a second time. It is cleared once the web
-        // view is actually back to full size instead.
-        this._iosHideSettle.arm(() => this._clearIosKeyboardBaseline());
-        this.document.body.classList.remove(BodyClass.isKeyboardVisible);
-        const root = this.document.documentElement;
-        this._setCssVar(root, CSS_VAR_KEYBOARD_HEIGHT, '0px');
-        this._setCssVar(root, CSS_VAR_KEYBOARD_OVERLAY_OFFSET, '0px');
-        this._updateIOSKeyboardViewportVars();
-      })
-      .then((handle) => this._keyboardListenerHandles.push(handle));
-
-    keyboard
-      .addListener('keyboardDidHide', () => this._iosHideSettle.run())
-      .then((handle) => this._keyboardListenerHandles.push(handle));
-
-    // Also handle focus changes while keyboard is already visible
-    this._focusinListener = (event: FocusEvent): void => {
-      const target = event.target as HTMLElement;
-      if (
-        this.document.body.classList.contains(BodyClass.isKeyboardVisible) &&
-        this._isInputElement(target)
-      ) {
-        // Small delay to let CSS padding apply, validate element is still focused
-        setTimeout(() => {
-          if (this.document.activeElement === target) {
-            this._scrollActiveInputIntoView();
-          }
-        }, 50);
-      }
-    };
-    this.document.addEventListener('focusin', this._focusinListener, { passive: true });
-
-    // Cleanup listeners on destroy
-    this._destroyRef.onDestroy(() => {
-      this._keyboardListenerHandles.forEach((handle) => handle.remove());
-      if (this._visualViewportResizeListener && window.visualViewport) {
-        window.visualViewport.removeEventListener(
-          'resize',
-          this._visualViewportResizeListener,
-        );
-      }
-      if (this._iosViewportChangeRaf !== null) {
-        window.cancelAnimationFrame(this._iosViewportChangeRaf);
-      }
-      this._iosShowSettle.cancel();
-      this._iosHideSettle.cancel();
-      if (this._focusinListener) {
-        this.document.removeEventListener('focusin', this._focusinListener);
-      }
-    });
-  }
-
-  /**
-   * The web view has finished growing back, so the next `keyboardWillShow` can
-   * take a fresh baseline — and must, in case the device rotated meanwhile.
-   */
-  private _clearIosKeyboardBaseline(): void {
-    this._iosViewportHeightBeforeKeyboard = 0;
-    this._updateIOSKeyboardViewportVars();
-  }
-
-  /**
-   * The web view has had its chance to resize around the keyboard, so everything
-   * frame-derived can act. Reached from `keyboardDidShow` or, if iOS drops it,
-   * from the fallback timer armed in `keyboardWillShow`.
-   */
-  private _settleIosKeyboard(): void {
-    this._isIosKeyboardSettled = true;
-    this._updateIOSKeyboardViewportVars();
-    // The shell height is a signal binding, so the shell is still at its
-    // pre-keyboard height until change detection runs. Measuring here would find
-    // the input comfortably inside a viewport that is about to shrink, and skip
-    // the scroll that keeps it off the keyboard (#9779).
-    afterNextRender(() => this._scrollActiveInputIntoView(), {
-      injector: this._environmentInjector,
-    });
-  }
-
-  private _updateIOSKeyboardViewportVars(): void {
-    const vars = computeIosKeyboardViewportVars({
-      keyboardHeight: this._iosKeyboardHeight,
-      baseHeight: this._iosViewportHeightBeforeKeyboard || window.innerHeight,
-      visualViewportHeight: window.visualViewport?.height,
-      isKeyboardSettled: this._isIosKeyboardSettled,
-      isKeyboardFrameUnreliable: this._iosKeyboardFrameUnreliable,
-    });
-
-    const root = this.document.documentElement;
-    let hasChanged = this._setCssVar(
-      // Not on <html>: a custom property there invalidates every element that
-      // could inherit it, which WebKit charges per node — measured at ~390ms per
-      // write on a 200-task list, on every frame of the keyboard animation
-      // (#9779). Only overlay panes read this one, so it lives on their
-      // container; the app shell gets a plain height below.
-      this._iosViewportVarTarget,
-      CSS_VAR_VISUAL_VIEWPORT_HEIGHT,
-      `${vars.visualViewportHeightPx}px`,
-    );
-    hasChanged =
-      this._setCssVar(
-        root,
-        CSS_VAR_KEYBOARD_OVERLAY_OFFSET,
-        `${vars.keyboardOverlayOffsetPx}px`,
-      ) || hasChanged;
-    if (vars.correctedKeyboardHeightPx !== null) {
-      hasChanged =
-        this._setCssVar(
-          root,
-          CSS_VAR_KEYBOARD_HEIGHT,
-          `${vars.correctedKeyboardHeightPx}px`,
-        ) || hasChanged;
-    }
-    this.iosShellHeight.set(
-      this._iosKeyboardHeight > 0
-        ? `calc(${vars.visualViewportHeightPx}px - var(--safe-area-top))`
-        : null,
-    );
-
-    // Every notification costs a synthetic window resize, and each of those makes
-    // CdkTextareaAutosize drop its caches and re-measure and every connected CDK
-    // overlay reposition — app-wide layout work. The keyboard animation fires a
-    // burst of visualViewport resize events, most of which land on values we have
-    // already written, so only tell the app about the ones that moved something.
-    if (hasChanged) {
-      this._notifyIOSViewportChange();
-    }
-  }
-
-  /**
-   * Writes a CSS variable, deduped; returns whether the value actually changed.
-   * Each variable has exactly one target, so the cache keys on the name alone.
-   */
-  private _setCssVar(target: HTMLElement, name: string, value: string): boolean {
-    if (this._cssVarCache.get(name) === value) {
-      return false;
-    }
-    this._cssVarCache.set(name, value);
-    target.style.setProperty(name, value);
-    return true;
-  }
-
-  private _notifyIOSViewportChange(): void {
-    if (this._iosViewportChangeRaf !== null) {
-      return;
-    }
-
-    this._iosViewportChangeRaf = window.requestAnimationFrame(() => {
-      this._iosViewportChangeRaf = null;
-      // Connected CDK overlays listen to viewport resize events via ViewportRuler.
-      window.dispatchEvent(new Event('resize'));
-    });
-  }
-
-  /**
    * Keyboard-height tracking via VisualViewport — the fallback path for any
    * non-iOS touch build (Capacitor Android, F-Droid, mobile-web).
    *
@@ -1019,63 +723,6 @@ export class GlobalThemeService {
         window.clearTimeout(resizeTimer);
       }
     });
-  }
-
-  private _isInputElement(el: HTMLElement): boolean {
-    const tagName = el.tagName.toLowerCase();
-    return (
-      tagName === 'input' ||
-      tagName === 'textarea' ||
-      tagName === 'select' ||
-      el.isContentEditable
-    );
-  }
-
-  /**
-   * Whether scrolling could bring this element into view at all.
-   *
-   * False only for an element sitting in a `position: fixed` container with no
-   * scrollable box in between — the global add-task bar, say: it moves with the
-   * viewport, so nothing can scroll it anywhere. An input inside a dialog keeps
-   * returning true, because the dialog's own scroll container can still lift it
-   * off the keyboard (#7388).
-   */
-  private _canScrollIntoView(el: HTMLElement): boolean {
-    let node: HTMLElement | null = el.parentElement;
-    while (node && node !== this.document.body) {
-      const style = window.getComputedStyle(node);
-      if (
-        node.scrollHeight > node.clientHeight &&
-        (style.overflowY === 'auto' || style.overflowY === 'scroll')
-      ) {
-        return true;
-      }
-      if (style.position === 'fixed') {
-        return false;
-      }
-      node = node.parentElement;
-    }
-    return true;
-  }
-
-  private _scrollActiveInputIntoView(): void {
-    const activeEl = this.document.activeElement as HTMLElement;
-    if (activeEl && this._isInputElement(activeEl)) {
-      // Mid keyboard animation the fixed add-task bar can sit outside the
-      // shrinking viewport just long enough for `scrollIntoViewIfNeeded` to
-      // scroll the list behind it to an arbitrary offset — which is one of the
-      // ways the bar appears to jump around (#9779). Scrolling cannot move a
-      // fixed element anyway, so skip it there.
-      if (!this._canScrollIntoView(activeEl)) {
-        return;
-      }
-      // scrollIntoViewIfNeeded is non-standard but well-supported in iOS WebView
-      if ('scrollIntoViewIfNeeded' in activeEl) {
-        (activeEl as any).scrollIntoViewIfNeeded(true);
-      } else {
-        activeEl.scrollIntoView({ behavior: 'smooth', block: 'center' });
-      }
-    }
   }
 
   /**
