+++ src/app/features/bottom-panel/bottom-panel-container.component.ts
@@ -455,7 +455,7 @@ export class BottomPanelContainerComponent implements AfterViewInit, OnDestroy {
       window.visualViewport.addEventListener('resize', this._boundOnViewportResize);
     }
     // Also the window event: on iOS the keyboard CSS variables can settle after
-    // the last visualViewport resize (GlobalThemeService only knows whether the
+    // the last visualViewport resize (IosKeyboardService only knows whether the
     // web view resized once `keyboardDidShow` fired, #9779) and it announces
     // that with a synthetic window resize. Without this the sheet would keep the
     // offset it computed mid-animation.
@@ -505,9 +505,9 @@ export class BottomPanelContainerComponent implements AfterViewInit, OnDestroy {
     if (!container) return;
 
     // Read the keyboard variables off the sheet itself, not <html>: the sheet is
-    // a CDK overlay, and --visual-viewport-height is set on the overlay
-    // container so writing it does not restyle the whole document (#9779).
-    // Inheritance makes this equivalent wherever a variable is set above.
+    // a CDK overlay, and on iOS all of them are set on the overlay container so
+    // writing them does not restyle the whole document (#9779). Inheritance
+    // makes this equivalent wherever a variable is set above.
     const keyboardStyle = getComputedStyle(container);
     const cssVisualViewportHeight = this._parseCssPx(
       keyboardStyle.getPropertyValue(CSS_VAR_VISUAL_VIEWPORT_HEIGHT),
