+++ src/app/features/tasks/add-task-bar/add-task-bar.component.spec.ts
@@ -28,6 +28,7 @@ import { getDbDateStr } from '../../../util/get-db-date-str';
 import { TaskRepeatCfgService } from '../../task-repeat-cfg/task-repeat-cfg.service';
 import { SS } from '../../../core/persistence/storage-keys.const';
 import { BodyClass } from '../../../app.constants';
+import { IosKeyboardService } from '../../../core/theme/ios-keyboard.service';
 import { IS_ANDROID_WEB_VIEW_TOKEN } from '../../../util/is-android-web-view';
 
 type ProjectServiceSignals = {
@@ -377,6 +378,22 @@ describe('AddTaskBarComponent', () => {
       fixture.nativeElement.style.setProperty('--keyboard-overlay-offset', '40px');
 
       expect(getComputedStyle(fixture.nativeElement).bottom).toBe('56px');
+    });
+
+    // The offset is bound on the host from IosKeyboardService rather than
+    // inherited from <html>, where writing it restyles the whole list (#9779).
+    it('takes the iOS keyboard offset from IosKeyboardService', () => {
+      document.body.classList.add(BodyClass.isIOS);
+      const iosKeyboardService = TestBed.inject(IosKeyboardService);
+
+      iosKeyboardService.keyboardOverlayOffset.set('40px');
+      fixture.detectChanges();
+      expect(getComputedStyle(fixture.nativeElement).bottom).toBe('56px');
+
+      // Hidden again: the host binding goes away and the :root default applies.
+      iosKeyboardService.keyboardOverlayOffset.set(null);
+      fixture.detectChanges();
+      expect(getComputedStyle(fixture.nativeElement).bottom).toBe('16px');
     });
 
     it('does not add the safe area to the keyboard offset for non-iOS touch builds', () => {
